// mobile/lib/types.ts
export type TimesheetStatus =
  | 'upcoming'
  | 'in_progress'
  | 'pending'
  | 'approved'
  | 'late'
  | 'discarded'
  | 'auto_closed'
  | 'paid';